Move TabbarItemType to types.js to remove cyclic import
