+++ packages/math-input/src/components/tabbar/types.js
@@ -1,0 +1,3 @@
+// @flow
+
+export type TabbarItemType = "Geometry" | "Operators" | "Numbers";
